Added componenets and form in the App.
