+++ src/App.js
@@ -1,35 +1,17 @@
 import React from 'react';
-
-/* Array to store a variable TodoList */
-const todoList = [
-  {
-    objectID: '0',
-    title: 'Complete assignment Lesson 1-1'
-  },
-  {
-    objectID: '1',
-    title: 'Upload it to GitHub'
-  },
-  {
-    objectID: '2',
-    title: 'Submit it to Squibby'
-  },
-];
+import TodoList from './TodoList';
+import AddTodoForm from './component/AddTodoForm';
 
 function App() {
   return (
-    <div style={{ textAlign: 'center' }}>
+    <div>
       <h1>Todo List</h1>
 
-      {/* Rendering todo list */}
-      <ul>
-        {todoList.map(function (listItem) {
-          return <li key={listItem.objectID}>
-              <span>{listItem.title}</span>
-            </li>
-          }
-        )}
-      </ul>
+      {/* Rendering todo list and the form */}
+      <AddTodoForm />
+      <TodoList />
+      
+
     </div>
   );
 }
