+++ src/component/AddTodoForm.js
@@ -1,0 +1,13 @@
+function AddTodoForm() {
+  return (
+    <form>
+      <div>
+        <label htmlFor="todoTitle">Title </label>
+        <input type="text" required id="todoTitle" />
+        <button>Add</button>
+      </div>
+    </form>
+  );
+}
+
+export default AddTodoForm;
